Moving GarbageCollectionNotificationInfo into its own public class
